Renamed BattleEngine.js file to index.js
