+++ src/engine/index.js
@@ -1,62 +1,86 @@
 import AI from '../AI'
 import Controls from '../controls'
 import Physics from '../physics'
-import SoundsManager from '../sounds/SoundsManager'
-import { soundsConfig } from '../sounds/SoundsConfig'
-import TexturesManager from '../textures/TexturesManager'
-import { texturesPaths, texturesConfig } from '../textures/TexturesConfig'
+
+import removeFrom from '../helpers/Array.remove'
+import InfoBars from '../infoBars'
+import Debugger from '../debug'
+import Dialog from '../effects/Dialog'
 
 
 // Engine contains all information about the world.
-export default class Engine {
-    constructor (game, width, height) {
+export default class BattleEngine {
+    constructor (game, level) {
         this.game = game
-        this.canvasWidth = width
-        this.canvasHeight = height
+        this.canvasWidth = game.canvasWidth
+        this.canvasHeight = game.canvasHeight
 
         this.AI = new AI(this, game)
         this.controls = new Controls(this, game)
         this.physics = new Physics(this, game)
-        this.soundsManager = new SoundsManager(this, game, soundsConfig)
-        this.texturesManager = new TexturesManager(this, game, texturesPaths, texturesConfig)
-    }
+        this.infoBars = new InfoBars(this, game)
 
-    init () {
-        this.hero = undefined
-        this.enemies = []
-        this.projectiles = []
-
-        // ?????
-        // this.roomWidth = width
-        // this.roomHeight = height
-        const width = this.canvasWidth
-        const height = this.canvasHeight
-
-        this.weapon1 = null
-        this.weapon2 = null
-
-        this.game.shots = 0
-        this.game.hits = 0
-
-        this.background = this.game.add.tileSprite(0, 0, width, height, 'sky')
-        this.game.world.setBounds(0, 0, height, width)
-
-        // Create Phaser groups for units and projectiles sprites (allUnitsSprites contains both heroes and enemies).
-        this.deadBodies = this.game.add.group()
-        this.allUnitsSprites = this.game.add.group()
-        this.enemySprites = this.game.add.group()
-        this.allUnitsSprites.addChild(this.enemySprites)
-        this.projectileSprites = this.game.add.group()
-
-        this.cinematicMode = false // If true - no unit can take actions.
+        this.debugger = new Debugger(this, game)
 
         this.TINT_RANGE = 80
         this.debugAttack = false
     }
 
-    loadResources () {
-        this.texturesManager.loadAssets()
-        this.soundsManager.loadSounds()
+    create () {
+        this.game.soundsManager.renderSounds()
+    }
+
+
+    render () {
+    }
+
+    // Initialize level of battle.
+    startLevel (LevelType) {
+        this.level = new LevelType(this)
+
+        this.worldWidth = this.level.width
+        this.worldHeight = this.level.height
+
+        this.initLevelData()
+        this.initObjects()
+        this.initWorldBounds(this.worldWidth, this.worldHeight)
+        this.initBackground(this.worldWidth, this.worldHeight, '')
+        this.initializePhaserSpriteGroups()
+        this.level.create()
+    }
+
+    initLevelData () {
+        this.shots = 0
+        this.hits = 0
+        this.score = 0
+        this.cinematicMode = false // If true - no unit can take actions.
+        this.timers = {}
+    }
+
+    initObjects () {
+        this.hero = undefined
+        this.enemies = []
+        this.projectiles = []
+        this.effects = []
+    }
+
+    // Set bound of the world.
+    initWorldBounds (worldWidth, worldHeight) {
+        this.game.world.setBounds(0, 0, worldWidth, worldHeight)
+    }
+
+    initBackground (width, height, texture) {
+        this.background = this.game.add.tileSprite(0, 0, width, height, texture)
+    }
+
+    // Create Phaser groups for units and projectiles sprites.
+    initializePhaserSpriteGroups () {
+        // !!!!!!!!!!!!!!!!!! First all previous groups must be removed - ADD THIS!
+        this.deadBodies = this.game.add.group()
+        this.allUnitsSprites = this.game.add.group() // contains both heroes and enemies
+        this.enemySprites = this.game.add.group()
+        this.allUnitsSprites.addChild(this.enemySprites)
+        this.projectileSprites = this.game.add.group()
     }
 
     // Level building.
@@ -64,50 +88,158 @@ export default class Engine {
         this.background.loadTexture(texture, 0)
     }
 
-    createUnit (x, y, UnitType) {
-        const newUnit = new UnitType(x, y)
-        this.physics.initPhysicsForUnit(newUnit)
-        return newUnit
+
+    cameraFollowUnit (unit) {
+        this.game.camera.follow(unit.sprite)
     }
 
-    createUnitAsHero (x, y, unitType) {
-        this.hero = this.createUnit(x, y, unitType)
-        this.hero.role = 'hero'
-        this.game.camera.follow(this.hero.sprite)
+    // cameraZoom () {
+    //     this.game.camera.shake(0.05, 500)
+    // }
+
+    cameraZoom (frames) {
+        let scale = 1.5
+        this.game.world.scale.setTo(scale, scale)
+        this.game.camera.shake(0.05, frames / 60 * 1000)
+        this.setTimer('camera', frames, () => {
+            this.game.world.scale.setTo(1, 1)
+        })
+    }
+    // , 100)
+        // })
+
+        
+
+    //     const zoomOut = function (x, game) {
+    //         var zoom = setInterval(() => {
+    //             x--;
+    //             game.world.scale.setTo(x, x);
+    //             if (x === 1 ) {
+    //                 game.camera.shake(0.05, 500);
+    //                 clearInterval(zoom);
+    //                 game.camera.follow(engine.hero.sprite);
+    //                 engine.cinematicMode = false;
+    //             };
+    //         }, 100);
+    //     }
+
+
+    createUnit (x, y, unitType) {
+        const unit = new unitType(this, x, y)
+        this.physics.initPhysicsForUnit(unit)
+        return unit
+    }
+
+    createHero (x, y, unitType) {
+        const hero = this.createUnit(x, y, unitType)
+        hero.role = 'hero'
+        this.hero = hero
+
+        // Add sprite to Phaser group.
+        this.allUnitsSprites.addChild(hero.sprite)
+        this.physics.initSpecialPhysicsForHeroUnit(hero)
+        return hero
+    }
+
+    createEnemy (x, y, unitType) {
+        const enemy = this.createUnit(x, y, unitType)
+        enemy.role = 'enemy'
+        this.enemies.push(enemy)
 
         //  Add sprite to Phaser group.
-        this.allUnitsSprites.addChild(this.hero.sprite)
-        this.physics.initSpecialPhysicsForHeroUnit(this.hero)
+        this.enemySprites.add(enemy.sprite)
+        return enemy
     }
 
-    createUnitAsEnemy (x, y, unitType) {
-        const newEnemy = this.createUnit(x, y, unitType)
-        newEnemy.role = 'enemy'
-        this.enemies.push(newEnemy)
+    // Boss has specific physics and AI.
+    createBoss (x, y, unitType) {
+        const boss = this.createUnit(x, y, unitType)
+        boss.role = 'boss'
+        this.boss = boss
+        this.enemies.push(boss)
 
         //  Add sprite to Phaser group.
-        this.enemySprites.add(newEnemy.sprite)
+        this.enemySprites.add(boss.sprite)
+        this.physics.initSpecialPhysicsForBossUnit(boss)
+        return boss
     }
 
-    // Boss is final boss with specific AI.
-    createUnitAsBoss (x, y, unitType) {
-        this.boss = this.createUnit(x, y, unitType)
-        this.boss.role = 'boss'
-        this.enemies.push(this.boss)
+    createWeapon (unit, weaponType) {
+        const weapon = new weaponType(this, unit)
+        return weapon
+    }
+
+
+    // Projectiles.
+    createProjectile (x, y, direction, projectileType, additional) {
+        const projectile = new projectileType(this, x, y, direction, additional)
+        this.projectiles.push(projectile)
 
         //  Add sprite to Phaser group.
-        this.enemySprites.add(this.boss.sprite)
-        this.physics.initSpecialPhysicsForBossUnit(this.boss)
-    }
+        this.projectileSprites.add(projectile.sprite)
+        this.physics.initPhysicsForProjectile(projectile)
+        return projectile
+    };
 
-    dealDamage (unit, damage) {
-        unit.hp -= damage
-        if (unit.hp <= 0) {
-            this.killUnit(unit)
+    removeProjectile (projectile) {
+        removeFrom(this.projectiles, projectile)
+        this.projectileSprites.remove(projectile.sprite)
+        projectile.sprite.kill()
+    };
+
+    setTimer (name, frames, callback) {
+        this.timers[name] = {
+            left: frames,
+            callback: callback
         }
     }
 
-    killUnit (unit) {
+
+    // Activity (Updates every frame).
+
+    // Here must be executed all activity of every object during current frame.
+    update () {
+        this.level.update()
+
+        // At first actions are setted, than actions are executed.
+        this.controls.setUnitActionsFromControls(this.hero)
+        this.hero.update()
+
+        this.enemies.forEach((enemy) => {
+            if (enemy.role === 'enemy') {
+                this.AI.setStrategyForUnit(enemy)
+            } else if (enemy.role === 'boss') {
+                this.AI.setStrategyForBoss(enemy)
+            }
+
+            this.AI.setActionsForUnit(enemy)
+            enemy.update()
+        })
+
+        this.projectiles.forEach((projectile) => {
+            projectile.update()
+        })
+
+        this.physics.handleProjectilesCollisions()
+        this.physics.handleUnitsCollisions()
+
+        this.effects.forEach((effect) => {
+            effect.update()
+        })
+
+        Object.keys(this.timers).forEach((name) => {
+            let timer = this.timers[name]
+            if (timer.left <= 0) {
+                timer.callback()
+                delete this.timers[name]
+            } else {
+                timer.left--
+            }
+        })
+    }
+
+
+    removeUnit (unit) {
         if (unit.role === 'hero') {
             this.allUnitsSprites.remove(unit.sprite)
             if (!unit.deathFrameCount || unit.deathFrameCount < 20) {
@@ -116,9 +248,9 @@ export default class Engine {
             }
         } else if (unit.role === 'enemy' || unit.role === 'boss') {
             this.createTheGuts(unit)
-            this.enemies.remove(unit)
+            removeFrom(this.enemies, unit)
             this.enemySprites.remove(unit.sprite)
-            this.scores += 1000
+            this.score += 1000
         }
         if (unit.weapon1 !== undefined) unit.weapon1.destroy()
         if (unit.weapon2 !== undefined) unit.weapon2.destroy()
@@ -134,7 +266,7 @@ export default class Engine {
         let sprite = unit.sprite
 
         if (!sprite.animations.getAnimation('death')) {
-            sprite = this.texturesManager.createSpriteByName(sprite.x, sprite.y, 'death_' + (unit.deathColor || 'green'))
+            sprite = this.game.texturesManager.createSpriteByName(sprite.x, sprite.y, 'death_' + (unit.deathColor || 'green'))
         }
 
         const scale = sprite.scale.x + sprite.scale.x * (Math.random() * 0.4 - 0.2)
@@ -154,202 +286,41 @@ export default class Engine {
         sprite.body && (sprite.body.moves = false)
     }
 
-    // Activity (Updates every frame).
 
-    // Here must be executed all activity of every object during current frame.
-    performActivityForObjects () {
-        // At first actions are setted, than actions are executed.
-        if (!this.cinematicMode) {
-            this.controls.setUnitActionsFromControls(this.hero)
-        } else {
-            this.hero.moving = false
-            this.hero.attacking1 = false
-            this.hero.attacking2 = false
-        }
 
-        this.performActivityForUnit(this.hero)
 
-        for (let i = this.enemies.length - 1; i >= 0; i--) {
-            const enemy = this.enemies[i]
-            if (enemy.role === 'enemy') {
-                this.AI.setStrategyForUnit(enemy)
-            } else if (enemy.role === 'boss') {
-                this.AI.setStrategyForBoss(enemy)
-            }
-
-            if (!this.cinematicMode) {
-                this.controls.setActionsForUnit(enemy)
-            } else {
-                enemy.moving = false
-                enemy.attacking1 = false
-                enemy.attacking2 = false
-            }
-
-            this.performActivityForUnit(enemy)
-        }
-
-        for (let i = this.projectiles.length - 1; i >= 0; i--) {
-            const projectile = this.projectiles[i]
-            this.performActivityForProjectile(projectile)
-        }
-
-        this.physics.handleProjectilesCollisions()
-        this.physics.handleUnitsCollisions()
-    }
-
-    performActivityForUnit (unit) {
+    // What happens when projectile hits unit.
+    handleProjectileHit (unit, projectile) {
+        unit.dealDamage(projectile.damage)
+        this.hits++
         if (unit.hp > 0) {
-            unit.modifyStats()
-            if (unit.role === 'boss') unit.update()
-            this.physics.performMovementForUnit(unit) // Implemented in physics.js module
-            unit.updateSprite()
-            if (unit.weapon1 !== undefined) {
-                unit.weapon1.isUsed = unit.attacking1
-                unit.weapon1.update()
-            }
-            if (unit.weapon2 !== undefined) {
-                unit.weapon2.isUsed = unit.attacking2
-                unit.weapon2.update()
-            }
-        }
+            this.removeProjectile(projectile)
+        };
+    };
+
+    win () {
+        this.game.score += this.score
+        this.displayScore = true
+        this.setTimer('nextLevel', 180, () => {
+            this.nextLevel()
+        })
     }
 
+    lose () {
+        new Dialog(this, this.hero, 'death-speech')
+        this.game.camera.fade('#000000', 1000)
+        this.game.soundsManager.stopSound('Soundtrack')
+        this.setTimer('death', 120, () => {
+            this.game.state.start('Death')
+        })
+    }
 
-    // // Attacks.
-    // // framesBeforeAttack indicates how many frames left to finish current attack.
-    // // If it is equal to 0 - current attack must be performed in current frame.
-    // // If it is lesser than 0 - weapon doesn't attack currently and can start new attack.
-    // //
-    // // framesAfterAttack indicates how many frames left to finisn 'cooldown' after attack.
-    // // If it is equal to 0 - new attack can be started.
-    // tryStartAttack (unit, weapon) {
-    //     if (weapon.framesBeforeAttack == -1 && weapon.framesAfterAttack == 0) {
-    //         // Attack can be started at this frame.
-
-    //         if (weapon.isShooting) {
-    //             this.tryStartShootingAttack(unit, weapon);
-    //         } else {
-    //             this.tryStartMeleeAttack(unit, weapon);
-    //         }
-
-    //     }
-    // }
-
-    // tryStartShootingAttack (unit, weapon) {
-    //     // Check ammo.
-    //     if (weapon.ammo > 0) {
-    //         weapon.ammo--;
-
-    //         weapon.framesBeforeAttack = weapon.delayBeforeAttack;
-    //         if (debugAttack) console.log("attack started"); // DEBUG
-    //     } else {
-    //         console.log("need to reload!");
-    //     }
-    // }
-
-    // tryStartMeleeAttack (unit, weapon) {
-    //     // Create melee special effect.
-    //     if (!weapon.isShooting) weapon.meleeEffect = new MeleeEffect(weapon);
-    //     weapon.framesBeforeAttack = weapon.delayBeforeAttack;
-    //     if (debugAttack) console.log("attack started"); // DEBUG
-    // }
-
-
-    // // Execute attack with weapon or decrease attacks 'cooldown'.
-    // performActivityForWeapon (weapon) {
-    //     // If weapon is attacking
-    //     if (weapon.framesBeforeAttack > -1) {
-    //         if (!weapon.isShooting) weapon.meleeEffect.update();
-    //         if (debugAttack && weapon.framesBeforeAttack > 0) console.log("attack in progress"); // DEBUG
-
-    //         // If attack must be performed now
-    //         if (weapon.framesBeforeAttack == 0) {
-    //             if (debugAttack) console.log("attack!"); // DEBUG
-    //             this.performAttack(weapon);
-    //         }
-    //         weapon.framesBeforeAttack--;
-    //     // If weapon is in 'cooldown' process.
-    //     } else if (weapon.framesAfterAttack > 0) {
-    //         if (debugAttack) console.log("can't attack"); // DEBUG
-    //         weapon.framesAfterAttack--;
-    //     } else if (weapon.framesToReload > -1) {
-    //         if (weapon.framesToReload == 0) {
-    //             console.log("reloaded!");
-    //         } else {
-    //             weapon.framesToReload--;
-    //         }
-    //     }
-
-    // };
-
-    // // Execute attack itself.
-    // performAttack (weapon) {
-    //     if (weapon.isShooting) {
-    //         this.shotFromWeapon(weapon, weapon.unit.direction);
-    //     } else {
-    //         this.hitWithMeleeWeapon(weapon);
-    //         weapon.meleeEffect.destroy();
-    //     }
-    //     // Launch the 'cooldown' process.
-    //     weapon.framesAfterAttack = weapon.delayAfterAttack;
-    // };
-
-    // // Perform shot from weapon.
-    // shotFromWeapon (weapon, direction) {
-    //     var shooter = weapon.unit;
-    //     var point = getOffsetPoint(shooter.sprite.x, shooter.sprite.y,
-    //         weapon.attackPoint.offset + shooter.direction, weapon.attackPoint.distance);
-    //     var additional = weapon.additional;
-    //     this.createProjectile(point.x, point.y, direction, weapon.projectile, additional);
-    // };
-
-    // // Perform melee weapon attack.
-    // hitWithMeleeWeapon (weapon) {
-    //     var attacker = weapon.unit;
-    //     var groupToCheck = attacker.role == 'hero' ? this.enemies : (attacker.role == 'enemy' ? [this.hero] : []);
-    //     var unitsHit = getHitUnits(weapon, groupToCheck);
-    //     unitsHit.forEach((unit) => {
-    //         this.dealDamage(unit, weapon.addDamage);
-    //     });
-    // };
-
-    // reloadWeapon (weapon) {
-    //     weapon.ammo = weapon.maxAmmo;
-    // }
-
-
-    // // Projectiles.
-    // createProjectile (x, y, direction, projectileType, additional) {
-    //     var newProjectile = new projectileType(x, y, direction, additional);
-    //     this.projectiles.push(newProjectile);
-
-    //     //  Add sprite to Phaser group.
-    //     this.projectileSprites.add(newProjectile.sprite);
-    //     initPhysicsForProjectile(newProjectile);
-    // };
-
-    // destroyProjectile (projectile) {
-    //     this.projectiles.remove(projectile);
-    //     this.projectileSprites.remove(projectile.sprite);
-    //     projectile.sprite.kill();
-    // };
-
-    // performActivityForProjectile (projectile) {
-    //     if (projectile.lifeTime <= 0) {
-    //         this.destroyProjectile(projectile);
-    //     } else {
-    //         projectile.lifeTime--;
-    //     };
-    //     performMovementForProjectile(projectile); // Implemented in physics.js module
-    // };
-
-    // handleProjectileHit (unit, projectile) {
-    //     this.dealDamage(unit, projectile.damage);
-    //     game.hits++;
-    //     if (unit.hp > 0) {
-    //         this.destroyProjectile(projectile);
-    //     };
-    // };
+    nextLevel () {
+        this.game.nextLevel()
+        if (this.game.getCurrentLevel()) {
+            this.game.state.start('Battle')
+        }
+    }
 
 
 }
